+++ lib/hooks/useUserCanCreateProposal.js
@@ -1,6 +1,6 @@
 import { ethers } from 'ethers'
 import { AuthControllerContext } from 'lib/components/contextProviders/AuthControllerContextProvider'
-import { useGovernorAlpha } from 'lib/hooks/useGovernanceData'
+import { useGovernorAlpha } from 'lib/hooks/useGovernorAlpha'
 import { useTokenHolder } from 'lib/hooks/useTokenHolder'
 import { useContext } from 'react'
